requirements.txt
audioread==3.0.1
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
contourpy==1.3.2
cycler==0.12.1
decorator==5.2.1
fonttools==4.58.4
idna==3.10
joblib==1.5.1
kiwisolver==1.4.8
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.44.0
matplotlib==3.10.3
msgpack==1.1.1
numba==0.61.2
numpy==2.2.6
packaging==25.0
pillow==11.2.1
platformdirs==4.3.8
pooch==1.8.2
pycparser==2.22
pyparsing==3.2.3
python-dateutil==2.9.0.post0
requests==2.32.4
scikit-learn==1.7.0
scipy==1.15.3
six==1.17.0
soundfile==0.13.1
soxr==0.5.0.post1
threadpoolctl==3.6.0
typing_extensions==4.14.0
urllib3==2.4.0
